+++ src/lte/test/lte-test-fdbet-ff-mac-scheduler.cc
@@ -519,7 +519,7 @@ LenaFdBetFfMacSchedulerTestCase2::DoRun()
      */
     for (int i = 0; i < m_nUser; i++)
     {
-        double thrRatio = (double)1 / m_nUser;
+        double thrRatio = 1.0 / m_nUser;
         double estThrRatio = (double)dlDataRxed.at(i) / totalData;
         NS_LOG_INFO("\tUser " << i << " thrRatio " << thrRatio << " estThrRatio " << estThrRatio);
         NS_TEST_ASSERT_MSG_EQ_TOL(estThrRatio, thrRatio, tolerance, " Unfair Throughput!");
