+++ src/wimax/model/ss-net-device.cc
@@ -1313,7 +1313,7 @@ SubscriberStationNetDevice::ProcessUcd(const Ucd& ucd)
         return; // nothing new in UCD so don't read
     }
     SetCurrentUcd(ucd);
-    m_linkManager->SetRangingCW((uint8_t)std::pow((double)2, (double)ucd.GetRangingBackoffStart()) -
+    m_linkManager->SetRangingCW((uint8_t)std::pow(2.0, (double)ucd.GetRangingBackoffStart()) -
                                 1); // initializing ranging CW
     OfdmUcdChannelEncodings ucdChnlEncodings = ucd.GetChannelEncodings();
